Moved the unit testing sample file into basics folder
