+++ Basics/MainClassTest.py
@@ -1,7 +1,11 @@
 import unittest
-import MainClass
 
+class MainClass:
+    def adding(self,num1,num2):
+        return num1+num2
 
+    def subtracting(self,num1,num2):
+        return num1-num2
 
 
 class MainClassTest(unittest.TestCase):
