--- UnitTesting/MainClass.py
@@ -1,7 +1,0 @@
-
-class MainClass:
-    def adding(self,num1,num2):
-        return num1+num2
-
-    def subtracting(self,num1,num2):
-        return num1-num2
